fix one more unsigned

+++ c++/src/Timezone.cc
@@ -749,7 +749,7 @@ namespace orc {
       variants[variant].gmtOffset =
         static_cast<int32_t>(decode32(ptr + variantOffset + 6 * variant));
       variants[variant].isDst = ptr[variantOffset + 6 * variant + 4] != 0;
-      unsigned nameStart = ptr[variantOffset + 6 * variant + 5];
+      uint64_t nameStart = ptr[variantOffset + 6 * variant + 5];
       if (nameStart >= nameCount) {
         std::stringstream buffer;
         buffer << "name out of range in variant " << variant
